rename type declaration file
