+++ src/components/organisms/TrainerMenu/index.tsx
@@ -1,7 +1,7 @@
 import React, { PureComponent } from 'react';
 import Card from '../../atoms/Card';
 import Flex from '../../atoms/Flex';
-import { Width, CaughtPokemon, PokeballIndex, PokeballSprites } from '../../..';
+import { Width, CaughtPokemon, PokeballIndex, PokeballSprites } from '../../../types';
 
 type Props = Width & {
   className: string;
