+++ src/store/actionCreator.ts
@@ -1,5 +1,5 @@
 import { CATCH_POKEMON } from './actions';
-import { Pokemon, Action } from '..';
+import { Pokemon, Action } from '../types';
 
 export function catchPokemon(pokemonObj: Pokemon): Action {
   return { type: CATCH_POKEMON, payload: pokemonObj };
